+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Version 1.0.0 - 04-Feb-2014
+
+- Rename package to `cviebrock/image-validator`
+- Require Laravel >= 4.1.18 due to validation replacement rules
+- Use better validation extension
+
+
 ## Version 0.1.1 - 16-Dec-2013
 
 - Initial release